backend/requirements.txt
acme==1.21.0
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.0
async-timeout==5.0.1
attrs==21.2.0
Automat==20.2.0
azure-common==1.1.28
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
Babel==2.8.0
bcrypt==3.2.0
billiard==4.2.1
blinker==1.4
boto3==1.40.23
botocore==1.40.23
cachetools==5.3.0
celery==5.5.3
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2025.8.3
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloud-init==23.3.3
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==3.0.9
django-filter==2.3.0
django-rest-swagger==2.2.0
django-storages==1.10.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et-xmlfile==1.1.0
exceptiongroup==1.3.0
future==0.15.2
google-api-core==1.34.0
google-api-python-client==1.12.1
google-auth==2.17.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.59.0
gunicorn==20.1.0
gyp==0.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
josepy==1.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kombu==5.5.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==6.0.1
MarkupSafe==2.1.2
mercadopago==2.3.0
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
openapi-codec==1.3.2
openpay==0.6
openpyxl==3.0.7
packaging==25.0
parsedatetime==2.6
pexpect==4.8.0
pillow==11.3.0
prompt_toolkit==3.0.52
protobuf==3.20.3
psycopg2==2.8.5
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.9.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.3
python-dotenv==1.1.1
python-magic==0.4.24
pytz==2021.1
PyYAML==5.3.1
redis==6.4.0
requests==2.32.5
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.9
s3transfer==0.13.1
SecretStorage==3.3.1
selectolax==0.4.1
sentry-sdk==1.39.1
service-identity==18.1.0
simplejson==3.19.1
six==1.17.0
sniffio==1.3.1
sos==4.5.6
sqlparse==0.5.3
ssh-import-id==5.11
systemd-python==234
Twisted==22.1.0
typing_extensions==4.14.1
tzdata==2025.2
ubuntu-advantage-tools==27.11.2
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==2.5.0
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
